change quantity field type

+++ app/core/models.py
@@ -97,7 +97,7 @@ class Ingredient(models.Model):
         on_delete=models.CASCADE,
         related_name='ingredients'
         )
-    quantity = models.DecimalField(max_digits=1000, decimal_places=2, default=0)
+    quantity = models.IntegerField(default=0)
     measurement = models.CharField(max_length=20, blank=True, null=True)
 
     class Meta:
